Refactored dashboard directory and fixed bugs
