+++ frontend/src/functionality/Dashboard/Dashboard.jsx
@@ -1,9 +1,9 @@
-import TaskGroup from './TaskGroup.jsx';
+import TaskGroup from './TaskSection/TaskGroup.jsx';
 import WelcomeMessage from '../../components/WelcomeMessage.jsx';
 import TaskSearchBar from './TaskSearchBar.jsx';
 import AddTaskButton from '../../components/AddTaskButton.jsx';
-import NotesSection from './NotesSection.jsx';
-import NoTasksMessage from './NoTasksMessage.jsx';
+import NotesSection from './NotesSection/NotesSection.jsx';
+import NoTasksMessage from './TaskSection/NoTasksMessage.jsx';
 import ErrorMessage from '../../components/ErrorMessage.jsx';
 import useDashboard from '../../utils/Hooks/useDashboard.js';
 import useBodyClass from '../../utils/Hooks/useBodyClass.js';
