+++ frontend/src/functionality/Dashboard/NotesSection/NotesTextArea.jsx
@@ -1,3 +1,5 @@
+import './stylesheets/NotesTextArea.css';
+
 /**
  * Renders the notes textarea input.
  *
