+++ frontend/src/functionality/Dashboard/TaskSection/TaskOptionsDropup/TaskOptionsDropup.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { FaEye, FaEdit, FaUndo, FaTrash } from 'react-icons/fa';
-import './stylesheets/TaskOptionsDropup.css';
+import '../../stylesheets/TaskOptionsDropup.css';
 
 /**
  * Drop-up menu for a task item with options to view details, edit, delete, and undo completion.
